fix: move HighlighterEdit to common.js for highlighColor
